Separate optional and required protobuf field decoding
