+++ crates/encoding/protobuf-encoding-components/src/impls/encode_mut/mod.rs
@@ -3,3 +3,5 @@ pub mod message;
 pub mod option;
 pub mod proto_field;
 pub mod proto_field_len;
+pub mod proto_field_required;
+pub mod proto_optional;
